module-info: use correct labels for jobs
The index in the array of jobs isn't the number!

Before this change the "Job #1" and "Job #2" were reversed.

Turns out, everywhere else we reference the jobs we always look up the
label in the labels map. That wasn't being done here and thus the jobs
were actually mis-labeled (2 was 1 and 1 was 2) making the chunk list
in module-info attributed to the wrong job.

Now it's fixed and the right job labels are show.